+++ src/graphql/getGraphqlTypeResolver.js
@@ -4,17 +4,15 @@ const castArrayShape = require('../utils/castArrayShape');
 const { walkmap } = require('../graph/traversal');
 const getGraphqlObjectType = require('./getGraphqlObjectType');
 
-// TODO: kill that bird
 function getGraphqlTypeResolver(g, iri) {
 
   // We look for all sub-classes of this rdfs:Class
   const classes = [...walkmap(g, iri, _rdfsSubClassOf)];
   const l = classes.length - 1;
 
-  return (source, info) => {
-    const types = g.resolvers.resolveSourceTypes(source, info);
-
-    // Everything is a resource, hence the fallback
+  return (source, info) => Promise.resolve(g.resolvers.resolveSourceTypes(source, info))
+  .then(types => {
+    // Everything is a rdfs:Resource, hence the fallback
     if (isNil(types)) return getGraphqlObjectType(g, rdfsResource);
 
     const typesArray = castArrayShape(types);
@@ -28,7 +26,7 @@ function getGraphqlTypeResolver(g, iri) {
 
     // This is the case when iri === rdfsResource
     return getGraphqlObjectType(g, typesArray[0]);
-  };
+  });
 }
 
 module.exports = getGraphqlTypeResolver;